✨ Feat: DiaryContent 이미지 provider 적용
