+++ src/pages/Diary/components/DiaryContent/DiaryContentDetail.tsx
@@ -1,5 +1,6 @@
 import DiaryContentImgs from '~/pages/Diary/components/DiaryContent/DiaryContentImgs';
 import DiaryContentText from '~/pages/Diary/components/DiaryContent/DiaryContentText';
+import { DiaryImgsProvider } from '~/pages/Diary/contexts/DiaryImgsContext';
 
 const DiaryContentDetail = () => {
   return (
@@ -8,7 +9,9 @@ const DiaryContentDetail = () => {
         다이어리 내용
       </span>
       <div className="rounded-xl border border-grey-200 px-3 py-5">
-        <DiaryContentImgs />
+        <DiaryImgsProvider>
+          <DiaryContentImgs />
+        </DiaryImgsProvider>
         <DiaryContentText />
       </div>
     </div>
