+++ src/pages/Diary/components/DiaryContent/DiaryImgsCarousel.tsx
@@ -1,0 +1,19 @@
+import { Carousel } from '~/components/domain';
+
+const DiaryImgsCarousel = () => {
+  const pictures = [
+    'https://picsum.photos/200',
+    'https://picsum.photos/200',
+    'https://picsum.photos/200',
+    'https://picsum.photos/200',
+    'https://picsum.photos/200',
+  ];
+
+  return (
+    <div className="h-[10rem]">
+      <Carousel pictures={pictures} />
+    </div>
+  );
+};
+
+export default DiaryImgsCarousel;
